Dynamic import for delta stream bundle
