+++ packages/drivers/odsp-driver/src/odspDocumentServiceFactoryWithCodeSplit.ts
@@ -24,7 +24,7 @@ export class OdspDocumentServiceFactoryWithCodeSplit
         super(
             getStorageToken,
             getWebsocketToken,
-            async () => import("./getSocketIo").then((m) => m.getSocketIo()),
+            async () => import(/* webpackChunkName: "socketModule" */ "./getSocketIo").then((m) => m.getSocketIo()),
             persistedCache,
             hostPolicy,
         );
