+++ packages/tools/webpack-fluid-loader/src/routes.ts
@@ -38,9 +38,9 @@ export function devServerConfig(baseDir: string, env: RouteOptions) {
             static: {
                 directory: path.join(
                     baseDir,
-                    "/node_modules/@fluid-tools/webpack-fluid-loader/dist/fluid-loader.bundle.js",
+                    "/node_modules/@fluid-tools/webpack-fluid-loader/dist/",
                 ),
-                publicPath: "/fluid-loader.bundle.js",
+                publicPath: "/code",
             },
             devMiddleware: {
                 publicPath: "/dist",
@@ -323,7 +323,7 @@ const fluid = (req: express.Request, res: express.Response, baseDir: string, opt
     <div id="content" style="min-height: 100%;">
     </div>
 
-    <script src="/fluid-loader.bundle.js"></script>
+    <script src="/code/fluid-loader.bundle.js"></script>
     ${umd.files.map((file) => `<script src="/${file}"></script>\n`)}
     <script>
         var pkgJson = ${JSON.stringify(packageJson)};
